main/requirements.txt
actionlib==1.14.0
angles==1.9.13
base-local-planner==1.17.3
bondpy==1.8.6
catkin==0.8.10
catkin-pkg==1.0.0
cv-bridge==1.16.2
diagnostic-updater==1.11.0
distro==1.9.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
gazebo-ros==2.9.3
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
image-geometry==1.16.2
interactive-markers==1.12.0
joint-state-publisher==1.15.1
laser-geometry==1.6.7
message-filters==1.17.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==1.24.4
opencv-contrib-python==4.11.0.86
pillow==10.4.0
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyopengltk==0.0.4
pyparsing==3.1.4
Pypubsub==4.0.3
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
PyYAML==6.0.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
resource-retriever==1.12.8
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospy==1.17.0
rosserial-python==0.9.2
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rtabmap-python==0.21.10
rviz==1.14.25
sensor-msgs==1.13.1
six==1.17.0
smclib==1.8.6
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tomli==2.2.1
topic-tools==1.17.0
turtlebot3-teleop==1.2.5
typing-extensions==4.13.0
xacro==1.14.19
